refactor: split custom entity files to one class per file
custom_selectors: entity.py → working_mode.py, select.py → solcast_likelihood.py
custom_switches: entity.py → description.py + switch.py
custom_times: entity.py → description.py + time.py
Updated imports in platform files and tests.
