+++ custom_components/hsem/custom_times/time.py
@@ -1,100 +1,24 @@
 """Time entity for the HSEM integration.
 
-Defines :class:`HSEMTimeEntityDescription` and :class:`HSEMTimeEntity`.
+Defines :class:`HSEMTimeEntity`.
 
 :class:`HSEMTimeEntity` is a standard :class:`TimeEntity` that persists its
 value to the config entry options so it survives HA restarts.
 """
 
-from dataclasses import dataclass
 from datetime import datetime, time
 from typing import Any
 
-from homeassistant.components.time import TimeEntity, TimeEntityDescription
+from homeassistant.components.time import TimeEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 
-from custom_components.hsem.entity import HSEMEntity
-from custom_components.hsem.utils.sensornames import (
-    get_ev_deadline_time_entity_id,
-    get_ev_deadline_time_key,
-    get_ev_deadline_time_unique_id,
-    get_ev_second_deadline_time_entity_id,
-    get_ev_second_deadline_time_key,
-    get_ev_second_deadline_time_unique_id,
-    get_schedule_1_end_time_entity_id,
-    get_schedule_1_end_time_key,
-    get_schedule_1_end_time_unique_id,
-    get_schedule_1_start_time_entity_id,
-    get_schedule_1_start_time_key,
-    get_schedule_1_start_time_unique_id,
-    get_schedule_2_end_time_entity_id,
-    get_schedule_2_end_time_key,
-    get_schedule_2_end_time_unique_id,
-    get_schedule_2_start_time_entity_id,
-    get_schedule_2_start_time_key,
-    get_schedule_2_start_time_unique_id,
-    get_schedule_3_end_time_entity_id,
-    get_schedule_3_end_time_key,
-    get_schedule_3_end_time_unique_id,
-    get_schedule_3_start_time_entity_id,
-    get_schedule_3_start_time_key,
-    get_schedule_3_start_time_unique_id,
+from custom_components.hsem.custom_times.description import (
+    _TIME_ID_MAP,
+    HSEMTimeEntityDescription,
 )
-
-# Map from config-entry key → (unique_id, entity_id)
-_TIME_ID_MAP: dict[str, tuple[str, str]] = {
-    get_schedule_1_start_time_key(): (
-        get_schedule_1_start_time_unique_id(),
-        get_schedule_1_start_time_entity_id(),
-    ),
-    get_schedule_1_end_time_key(): (
-        get_schedule_1_end_time_unique_id(),
-        get_schedule_1_end_time_entity_id(),
-    ),
-    get_schedule_2_start_time_key(): (
-        get_schedule_2_start_time_unique_id(),
-        get_schedule_2_start_time_entity_id(),
-    ),
-    get_schedule_2_end_time_key(): (
-        get_schedule_2_end_time_unique_id(),
-        get_schedule_2_end_time_entity_id(),
-    ),
-    get_schedule_3_start_time_key(): (
-        get_schedule_3_start_time_unique_id(),
-        get_schedule_3_start_time_entity_id(),
-    ),
-    get_schedule_3_end_time_key(): (
-        get_schedule_3_end_time_unique_id(),
-        get_schedule_3_end_time_entity_id(),
-    ),
-    get_ev_deadline_time_key(): (
-        get_ev_deadline_time_unique_id(),
-        get_ev_deadline_time_entity_id(),
-    ),
-    get_ev_second_deadline_time_key(): (
-        get_ev_second_deadline_time_unique_id(),
-        get_ev_second_deadline_time_entity_id(),
-    ),
-}
-
-
-@dataclass(frozen=True)
-class HSEMTimeEntityDescription(TimeEntityDescription):
-    """Extended entity description that adds a human-readable description field.
-
-    Attributes
-    ----------
-    description:
-        Short human-readable description of the time entity's purpose, exposed
-        as an entity attribute for dashboard display.
-    default_value:
-        Initial time value as an ``"HH:MM:SS"`` string.
-    """
-
-    description: str = ""
-    default_value: str = "00:00:00"
+from custom_components.hsem.entity import HSEMEntity
 
 
 class HSEMTimeEntity(TimeEntity, HSEMEntity):
